includes= in library.properties
from https://github.com/arduino/arduino/wiki/arduino-ide-1.5:-library-specification:

includes - (available from IDE 1.6.10) (optional) a comma separated list of files to be added to the sketch as #include <...> lines. This property is used with the "Include library" command in the IDE. If the includes property is missing all the headers files (.h) on the root source folder are included.